pyproject.toml
[tool.poetry]
name = "audio-files"
version = "0.1.0"
description = ""
authors = ["Сократ"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
dishka = "^1.5.1"
sqlalchemy = {version = "^2.0.40", extras = ["asyncio"]}
orjson = "^3.10.16"
alembic = "^1.15.2"
asyncpg = "^0.30.0"
python-jose = "^3.4.0"
aiohttp = "^3.11.14"
aiofiles = "^24.1.0"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
